fix: chain dual-branch interactions instead of orphaning; centralized env vars to env.py; Remove convenience getters; all consumers use load_env()
